update to new source code[beetp]

+++ src/main/java/org/stone/beetp/BeeTaskPoolMonitorVo.java
@@ -25,6 +25,6 @@ public interface BeeTaskPoolMonitorVo {
 
     int getTaskRunningCount();
 
-    int getTaskCompletedCount();
+    long getTaskCompletedCount();
 
 }
